pyproject.toml
[tool.poetry]
name = "klembord"
version = "0.1.5"
description = "Full toolkit agnostic cross-platform clipboard access"
readme = "README.md"
keywords = ["clipboard", "cross-platform"]
authors = ["Ozymandias <tomas.rav@gmail.com>"]
license = "MIT"
repository = "https://github.com/OzymandiasTheGreat/klembord"
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Environment :: X11 Applications",
	"Environment :: Win32 (MS Windows)",
	"Intended Audience :: Developers",
	"Operating System :: Microsoft :: Windows",
	"Operating System :: POSIX",
	"Operating System :: POSIX :: Linux",
]

[tool.poetry.dependencies]
python = "^3.4"
python-xlib = { version = "^0.26", markers = "sys_platform == 'linux'" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
